Completely remove company
Company really never worked 100% for me.
I wanted to have preview+completion-list separated. the fist automatic,
and the second one working only on demand... That was hard with company
and required a very extensive config hacking stuff.

I already tried corfu in the past, but it didn't work in tui,
But now there is a new corfu-terminal, which seems to make it work
totally functional for me.
And emacs added the completion-preview-mode both of them use the capf
and native stuff. So I will go for them